# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.changePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        type: string
    required:
    - current_password
    - new_password
    type: object
...
  handler.createPostRequest:
    properties:
      content:
        type: string
      cover_image_url:
        type: string
      excerpt:
        type: string
      slug:
        type: string
      tag_ids:
        items:
          type: string
        type: array
      title:
        type: string
    required:
    - title
    type: object
...
  handler.errorResponse:
    properties:
      code:
        type: string
      error:
        type: string
    type: object
...
  handler.togglePublishRequest:
    properties:
      published:
        type: boolean
    type: object
...
  handler.updateMeRequest:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      display_name:
        type: string
    type: object
  handler.updatePostRequest:
    properties:
      content:
        type: string
      cover_image_url:
        type: string
      excerpt:
        type: string
      slug:
        type: string
      tag_ids:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: A production-ready personal blog REST API
  title: Simple Blog API
  version: "1.0"
paths:
...
  /me:
    get:
      description: Returns the authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - profile
    patch:
      consumes:
      - application/json
      description: Updates the authenticated user's profile
      parameters:
      - description: Profile data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.updateMeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Update current user profile
      tags:
      - profile
  /me/change-password:
    post:
      consumes:
      - application/json
      description: Changes the authenticated user's password
      parameters:
      - description: Current and new password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.changePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - profile
  /posts:
    get:
      description: Returns a paginated list of blog posts
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search query
        in: query
        name: q
        type: string
      - description: Filter by tag
        in: query
        name: tag
        type: string
      - description: Filter by author ID
        in: query
        name: author
        type: string
      - description: Sort order
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: List posts
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: Creates a new blog post
      parameters:
      - description: Post data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.createPostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Create a post
      tags:
      - posts
  /posts/{id}:
    delete:
      description: Deletes a blog post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete a post
      tags:
      - posts
    get:
      description: Returns a single blog post by slug or UUID
      parameters:
      - description: Post slug or UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Get a post
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: Updates an existing blog post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - description: Post data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.updatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Update a post
      tags:
      - posts
...
  /posts/{id}/publish:
    patch:
      consumes:
      - application/json
      description: Publishes or unpublishes a blog post
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - description: Publish flag
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.togglePublishRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Toggle post publish status
      tags:
      - posts
...
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"